backend/requirements.txt
Flask==1.1.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
Flask-RESTful==0.3.8
python-dotenv==0.11.0
flask-cors
pytest==5.4.1
python-dateutil
